# CKZS_API/docs/ckzs-api.openapi.yaml
openapi: 3.1.0
info:
  title: CKZS API
  version: 1.0.0
  description: CKZS App 与管理端后端接口定义
servers:
  - url: /
    description: 当前服务
...
paths:
...
  /api/devices/batchCommand:
    post:
      tags: [设备]
      x-apifox-folder: 设备
      operationId: batchCommand
      summary: 批量下发设备指令
      security: [{ bearerAuth: [] }]
      requestBody:
        $ref: '#/components/requestBodies/BatchCommandRequest'
      responses:
        '200': { $ref: '#/components/responses/JsonResponse' }
  /api/devices/batchQueryStatus:
    post:
      tags: [设备]
      x-apifox-folder: 设备
      operationId: batchQueryStatus
      summary: 批量查询设备状态
      security: [{ bearerAuth: [] }]
      requestBody:
        $ref: '#/components/requestBodies/BatchQueryStatusRequest'
      responses:
        '200': { $ref: '#/components/responses/JsonResponse' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  requestBodies:
...
    BatchCommandRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [deviceCodes, type]
            properties:
              deviceCodes:
                type: array
                minItems: 1
                maxItems: 5000
                uniqueItems: true
                items: { type: string }
              type: { type: string }
              params:
                type: object
                additionalProperties: true
    BatchQueryStatusRequest:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [deviceCodes]
            properties:
              deviceCodes:
                type: array
                minItems: 1
                maxItems: 10
                uniqueItems: true
                items: { type: string }
              timeout: { type: integer, minimum: 1000, maximum: 10000, default: 5000 }
...
  responses:
    JsonResponse:
      description: 统一 JSON 响应
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AnyObject'
  schemas:
    AnyObject:
      type: object
      additionalProperties: true
...